Change orderValue type to Double

+++ src/main/scala/streaming/StreamingEntry.scala
@@ -7,12 +7,12 @@ case class StreamingEntry(id: String = "abide",
                          name: String = "John",
                          surname: String = "Lennon",
                          numberOfCalls: Int = 0,
-                         totalSales: Int = 0) {
+                         totalSales: Double = 0.0) {
   override def toString: String = {
     this.getClass.getDeclaredFields.map(_.get(this)).mkString(",")
   }
 
-  def toTuple: (String, String, String, Int, Int) = {
+  def toTuple: (String, String, String, Int, Double) = {
     (id, name, surname, numberOfCalls, totalSales)
   }
 }
@@ -44,6 +44,6 @@ object StreamingEntry {
 
   def parseFromString(str: String): StreamingEntry = {
     val fields = str.split(",")
-    StreamingEntry(fields(0), fields(1), fields(2), fields(3).toInt, fields(4).toInt)
+    StreamingEntry(fields(0), fields(1), fields(2), fields(3).toInt, fields(4).toDouble)
   }
 }
